support creation of aws clients in a blocking context (#579)
Esthri supports calling some function in a blocking context using the
`#[tokio::main]` macro. Unfortunately with the move to the aws_sdk
crate, the new aws client can only be created in an async function, sort
of completely destroying the point of the blocking.rs module.

Here I added the ability to create a default client in a blocking
context, so one can use blocking.rs modules functions in an e2e scenario
in a different crate.

In addition, I want to be able to specify the client's region, so I
added that to the s3 client initiator.